requirements.txt
PyQt6>=6.6.0
psycopg2-binary>=2.9.9
psutil>=5.9.0
pywin32>=306; sys_platform == 'win32'
pandas>=2.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
